[Firebase][MLKit] Enhanced binding
